Allow permissions to be functions (#278)
API changes
====

`modelDefinition.permissions.crudOperations` can be defined both as `booleans` and as `functions`. 

`edit` and `delete` permission functions should be able to act in 2 manners: `global` and `scoped to instance`. If function is called without arguments -> this is a request for `global` permissions (editor-wide, not connected to `instance`/`business object`). If a function is called with argument - it is an object with `instance` property, and it is used to define permissions for this particular object.

Changes in `adjacent item` navigation: no matter which view a user is on, upon click on `next`/`prev` buttons editor navigates to `show` or `edit` view according to permissions for the instance. If instance is editable - it opens in `edit` view, otherwise `show`.